cityzen_backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
whitenoise
python-docx
django-filter
drf-spectacular
Pillow
channels
daphne
pypdf